# swagger.yaml
openapi: 3.0.3
info:
  title: Vulnerable Auction Lab API
  description: |
    REST API for the Vulnerable Auction Lab — an educational microservices auction platform
    with **intentional security vulnerabilities** (SQL injection, XSS, weak JWT, missing auth
    checks, IDOR, permissive CORS). These vulnerabilities exist for security training purposes.
  version: 1.0.0
  contact:
    name: Auction Lab
servers:
  - url: https://hkku6t3qsd.execute-api.us-east-1.amazonaws.com/dev/api
    description: Production (AWS Lambda)
  - url: http://localhost:3001/api
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from /auth/login. Weak secret and no expiration (intentional vulnerability).
        Token can also be passed in the request body as `token` field.
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        password:
          type: string
          description: Bcrypt hash — intentionally exposed (vulnerability)
        name:
          type: string
        role:
          type: string
          enum: [user, admin, seller]
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    Auction:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        starting_price:
          type: number
          format: decimal
        current_bid:
          type: number
          format: decimal
        end_time:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, ended, cancelled]
        workflow_state:
          type: string
          enum: [active, pending_sale, shipping, complete]
        created_by:
          type: integer
          description: User ID of creator
        winner_id:
          type: integer
          nullable: true
        closed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
...
    Bid:
      type: object
      properties:
        id:
          type: integer
        auction_id:
          type: integer
        user_id:
          type: integer
        amount:
          type: number
          format: decimal
        payment_status:
          type: string
          default: pending
        created_at:
          type: string
          format: date-time
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List all users
      description: |
        Returns all users including password hashes.
        **Vulnerabilities:** no admin check, password hashes exposed.
      operationId: listUsers
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      description: |
        Returns user with bids and auctions. **Vulnerability:** IDOR, password hash exposed.
      operationId: getUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/User'
                  - type: object
                    properties:
                      bids:
                        type: array
                        items:
                          $ref: '#/components/schemas/Bid'
                      auctions:
                        type: array
                        items:
                          $ref: '#/components/schemas/Auction'
        '404':
          description: User not found

    put:
      tags: [Users]
      summary: Update user
      description: |
        Updates any user fields. If password provided, it's hashed with bcrypt (rounds=5).
        **Vulnerabilities:** no ownership check (IDOR), weak bcrypt, password hash returned.
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                role:
                  type: string
                phone:
                  type: string
                address:
                  type: string
              additionalProperties: true
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: No fields to update
        '404':
          description: User not found
...